refactor: User 엔티티 내 중복된 @Builder.Default 제거 및 KakaoLoginResponseDTO 파일 위치 변경
